requirements.txt
requests~=2.24.0
validators~=0.18.1
python-dotenv~=0.14.0